Issue #23 Cleaned up Doxygen page issues.
Fixed error in release doxygen links (all packages were linked to
Amesos documentation). Removed obsolete links to Doxygen collectsions
that no longer exist. Added links to release 12.16 Doxygen, which
is the most recently available release version.

On the server, I soft linked "latest-release" to the 12.16 version.
This can be updated when appropriate to always point to the latest.